Improve logout behaviour
A recent AIMS update appears to use much longer server sessions. This
causes login to fail if a new connection is made, including from the
crew portal. While we already check for this login failure and carry
out a logout as necessary, this commit ensures that aims-extract
always logs out.