Default to using ecrew.easyjet.com rather than connected portal.
A new URL has become available to access AIMS that does not involve
routing through the connected portal, and is thus much quicker to log
on to. This commit connects using this new URL, with connecting via
the crew portal an option with a -p or --portal command line switch.